test1

+++ 1/1.c
@@ -2,13 +2,14 @@
 #include <math.h>
 #define EPSILON 1e-7
 
+
 double bisection(int p, int q, double (*func)(int, int, double));
 double f(int p, int q, double x);
 int main() {
     int p;
     int q;
     scanf("%d%d", &p, &q);
-    printf("%.4f\n", bisection(p, q, f));
+    printf("%.4lf\n", bisection(p, q, f));
     return 0;
 }
 
@@ -17,7 +18,7 @@ double bisection(int p, int q, double (*func)(int, int, double)) {
     double b = 20;
     double xp;
     xp = (a + b) / 2;
-    int c = func(p, q , xp);
+    double c = func(p, q , xp);
         while(fabs(c) >= EPSILON)
         {
             if((c * func(p, q , a)) > 0)
